+++ RELEASE_NOTES.md
@@ -1,6 +1,13 @@
 ## RELEASE NOTES
 
 
+### Version 0.3.2 - Mar 3, 2014
+
+**Fixes**
+- DocumentParser.getSearchNumberField(): NullPointerException
+- ObjectParser.getDocumentFieldValue(): java.lang.IllegalArgumentException - if the document does not have exactly one field with the name
+
+
 ### Version 0.3.0 - Mar 1, 2014
 
 **New**
@@ -10,12 +17,10 @@
 - Refactoring of the class ObjectParser
 
 
-
 ### Version 0.2.0 - Feb 19, 2014
 
 **New**
 - Improvement to permit types String, Long or Integer for DocumentId fields
-
 
 
 ### Version 0.1.0 - Feb 11, 2014